requirements.txt
Django>=4.2
djangorestframework>=3.14
mysqlclient>=2.1
python-dotenv>=1.0.0
gunicorn>=20.1.0
django-cors-headers>=4.0
